Update README.md
TComPortDrv v3.1.2

This is a Unicode and further improved version of the CpDrv non-visible component for communicating via RS232 COM-Ports on Windows. The code is published under the Apache 2.0 license with the consent of its original author. Due to the use of $(AUTO) for the libsuffix it is compatible from D10.4.1 onwards only. If you remove this setting in the package project option of both packages it should work with older versions as well. The code found in SerialPorts.pas is not working yet.

[Download TComPortDrv v3.1.2](https://developer.team/delphi/35321-tcomportdrv-v312.html)